Updated to Fedora 36

+++ minimal-plasma-installation/minimal-plasma-installation.sh
@@ -7,7 +7,7 @@
 # Author: Bastien 'locobastos' MARTIN                                          #
 # Licence: Do what you want with this script, but please contact me if you     #
 #          improve it, I may be interessted on your enhancements.              #
-# Date: 20/09/2019                                                             #
+# Date: 19/08/2022                                                             #
 ################################################################################
 
 # Check if the current user is root
@@ -18,14 +18,31 @@ then
 fi
 
 # Update your OS
-dnf upgrade -y
+if ! dnf upgrade --assumeyes
+then
+    echo "Error while upgrading your OS."
+    exit
+fi
 
-# Install X.Org & Plasma Desktop
-dnf install -y plasma-desktop sddm sddm-kcm xorg-x11-utils xdg-user-dirs
+# Install X.Org & PLASMA Desktop
+if ! dnf install --assumeyes kscreen plasma-desktop sddm sddm-kcm xdg-user-dirs
+then
+    echo "Error while installing packages."
+    exit
+fi
 
 # Enable SDDM
-systemctl enable sddm graphical.target
-systemctl set-default graphical.target
+if ! systemctl enable sddm graphical.target
+then
+    echo "Error while enabling SDDM."
+    exit
+fi
+
+if ! systemctl set-default graphical.target
+then
+    echo "Error while enabling SDDM."
+    exit
+fi
 
 # Reboot
 reboot
